+++ 1.x/Source/ddownloadwindowitem.cc
@@ -417,8 +417,8 @@ void ddownloadwindowitem::updateProgress(const qint64 done,
     {
       QFileInfo fileInfo(m_dstFileName);
       int completed =
-	100 * static_cast<double> (fileInfo.size()) /
-	qMax(1.0, static_cast<double> (m_total));
+	static_cast<int> (100.0 * static_cast<double> (fileInfo.size()) /
+			  qMax(1.0, static_cast<double> (m_total)));
 
       if(completed <= ui.progressBar->maximum())
 	ui.progressBar->setValue(completed);
@@ -469,10 +469,10 @@ void ddownloadwindowitem::updateProgress(const qint64 done,
 		      static_cast<long double> (m_rate) +
 		      static_cast<long double> (currentSize - m_lastSize) /
 		      static_cast<long double> (secs)) >= 1.0)
-	      m_rate =
-		static_cast<long double> (currentSize -
-					  m_lastSize) /
-		static_cast<long double> (secs);
+	      m_rate = static_cast<qint64>
+		(static_cast<long double> (currentSize -
+					   m_lastSize) /
+		 static_cast<long double> (secs));
 
 	  if(m_total > 0 && m_total >= fileInfo.size())
 	    {
@@ -924,7 +924,7 @@ void ddownloadwindowitem::slotComputeFileHash(void)
       qint64 rc = 0;
 
       while((rc = file.read(buffer.data(), buffer.length())) > 0)
-	hash.addData(buffer, rc);
+	hash.addData(buffer, static_cast<int> (rc));
 
       file.close();
       ui.fileHash->setText(hash.result().toHex());
