+++ 1.x/Source/dtabwidget.cc
@@ -487,7 +487,8 @@ void dtabwidget::setTabButton(int index)
 #endif
 
       widget->setMaximumHeight
-	(multiplier * qMax(label->height(), progressBar->height()));
+	(static_cast<int>
+	 (multiplier * qMax(label->height(), progressBar->height())));
     }
 
   m_tabBar->setTabButton(index, side, 0);
